refactor: combine tls tests to single file
